Minor Cleanup of imports and user guide url
git-svn-id: http://webgoat.googlecode.com/svn/trunk@105 4033779f-a91e-0410-96ef-6bf7bf53c507

+++  webgoat/main/project/JavaSource/org/owasp/webgoat/lessons/NewLesson.java
@@ -67,7 +67,7 @@ public class NewLesson extends LessonAdapter
     	return GENERAL;
     }
 
-    private final static Integer DEFAULT_RANKING = new Integer(5);
+    private final static Integer DEFAULT_RANKING = new Integer(85);
 
 
     protected Integer getDefaultRanking()
